LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/intrinsic-sizing/grid-lanes-baseline-alignment-affects-intrinsic-size-006-expected.txt

É
É
É

FAIL .grid-lanes 1 assert_equals:
<div class="grid-lanes" style="grid-template-rows: minmax(0px, auto) minmax(0px, auto);" data-expected-height="120">
  <!-- plum: span-2, padding-top: 40px raises its baseline, setting the shared baseline. -->
  <div style="font-size: 20px; padding-top: 40px; width: 60px; background: plum; grid-row: 1/3;">É</div>
  <!-- lightblue: span-1 in row 1, gets a 40px shim to align with the shared baseline. -->
  <div style="font-size: 20px; width: 60px; background: lightblue; grid-row: 1;">É</div>
  <!-- lightgreen: span-1 in row 2, also gets a 40px shim. -->
  <div style="font-size: 20px; width: 60px; background: lightgreen; grid-row: 2;">É</div>
  <!-- lightyellow: span-1 in row 2, height: 100% resolves against the track size (60px). -->
  <div style="font-size: 20px; width: 60px; height: 100%; background: lightyellow; grid-row: 2; align-self: normal" data-expected-height="60"></div>
</div>
height expected 120 but got 80
